20.07 commit

+++ backend/controllers/PromotionController.php
@@ -4,6 +4,7 @@ namespace backend\controllers;
 
 use backend\models\PromotionForm;
 use common\models\Promotion;
+use service\PromotionService;
 use Yii;
 use yii\data\ActiveDataProvider;
 use yii\filters\AccessControl;
@@ -12,6 +13,13 @@ use yii\web\UploadedFile;
 
 class PromotionController extends  Controller
 {
+    // public PromotionService $promotion_service;
+    // public function __construct($id, $module, PromotionService $promotion_service, $config = [])
+    // {
+    //     $this->promotion_service = $promotion_service;
+    //     parent::__construct($config);
+    // }
+
     /**
      * {@inheritdoc}
      */
@@ -68,7 +76,12 @@ class PromotionController extends  Controller
             }
         }
 
-        return $this->render('create', ['model' => $model]);
+        return $this->render('create', [
+            'model' => $model,
+            'initial_prev' => [],
+            'image_conf' => [],
+            'promotion_id' => ''
+        ]);
     }
 
     public function actionDelete($id)
@@ -77,7 +90,8 @@ class PromotionController extends  Controller
         if ($promotion) {
             $url_image = $promotion->url_image;
             if ($promotion->delete()) {
-                unlink($url_image);
+                if (file_exists($url_image)) unlink($url_image);
+
                 Yii::$app->session->addFlash('success', 'Акция удалена');
             } else {
                 Yii::$app->session->addFlash('error', 'Ошибка удаления акции');
@@ -90,23 +104,62 @@ class PromotionController extends  Controller
 
     public function actionUpdate($id)
     {
-        echo 'actionUpdate';
-        // return $this->render('edit');
+        $promotion = Promotion::findOne(['id' => $id]);
+        $model = new PromotionForm();
+
+        if ($model->load(Yii::$app->request->post())) {
+            // $this->promotion_service->saveDB($model);
+
+            $model->imageFile = UploadedFile::getInstance($model, 'imageFile');
+
+            if ($url_image = $model->upload()) {
+                $promotion->name = $model->name;
+                $promotion->description = $model->description;
+                // die(var_dump($url_image));
+                if ($url_image != $promotion->url_image) {
+                    if (file_exists($promotion->url_image)) unlink($promotion->url_image);
+                }
+                $promotion->url_image = $url_image;
+                if ($promotion->save()) {
+                    Yii::$app->session->addFlash('success', 'Акция сохранена');
+                } else {
+                    Yii::$app->session->addFlash('error', 'Ошибка сохранения акции');
+                }
+                return $this->redirect(['/promotion/index']);
+            }
+        }
+
+        $model->name = $promotion->name;
+        $model->description = $promotion->description;
+        $initial_prev = ['../../' . $promotion->url_image];
+        $name_file = explode('/', $promotion->url_image);
+        $image_conf = [
+            [
+                'key' => $name_file[count($name_file) - 1],
+            ]
+        ];
+
+        return $this->render('create', [
+            'model' => $model,
+            'initial_prev' => $initial_prev,
+            'image_conf' => $image_conf,
+            'promotion_id' => $id
+        ]);
     }
 
     public function actionView($id)
     {
         $promotion = Promotion::findOne(['id' => $id]);
+        $model = new PromotionForm();
         if ($promotion) {
+            $imageFile = $model->imageFile;
             // die(var_dump($promotion));
-            return $this->render('view', ['promotion' => $promotion]);
+            return $this->render('view', [
+                'model' => $model,
+                'promotion' => $promotion
+            ]);
         } else {
             Yii::$app->session->addFlash('error', 'Акция не найдена');
         }
-    }
-
-    public function actionEdit()
-    {
-        return $this->render('edit');
     }
 }
